HTBHF-2391 Refactor out common code for payment cycle integration tests.
